backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
motor==3.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
python-socketio==5.10.0
alembic==1.12.1
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
pillow==10.1.0

